+++ frontend/src/components/product/reviewsTab.tsx
@@ -2,7 +2,7 @@ import ReviewStar from "./ReviewsStar";
 import ReviewPopup from "./ReviewWriteModal";
 import ReviewDetailModal from "./ReviewDetailModal";
 
-import { PaginationDemo } from "../common/Paginaionm";
+import { PaginationDemo } from "../common/Paginaion";
 import Reviews from "./Reviews";
